refactor: the Integrations catalog belongs to the cluster, not to Settings
It already had its own route; the code, the search machinery and the
copy still lived under settings. Moved beside the cluster's own screens,
with plain groups instead of the settings search, and
/settings/integrations is no longer a special case.
