+++ src/pages/IntegrationsList.tsx
@@ -1,21 +1,12 @@
 /**
- * The Integrations catalog, at its own door.
+ * The Integrations catalog: what this cluster has.
  *
- * It used to be a Settings section, and that shirt never fit: what a
- * cluster has is not a preference of this app's. A theme belongs to the
- * reader; cert-manager belongs to the cluster. Splitting the doors is what
- * lets the sidebar say both things plainly — Settings holds decisions,
- * this page holds an inventory.
- *
- * The content is the same one screen allowed to name an extension, so the
- * search context it indexes itself under comes along with it.
+ * A cluster's page, not a preference of this app's. A theme belongs to the
+ * reader; cert-manager belongs to the cluster, which is why this lives
+ * under the cluster's routes beside its workloads and not in Settings.
  */
 
-import { IntegrationsSettings } from "@/components/settings/IntegrationsSettings";
-import {
-  SettingsSearchProvider,
-  SettingsSectionScope,
-} from "@/components/settings/settings-search";
+import { IntegrationsCatalog } from "@/components/cluster/IntegrationsCatalog";
 import { SectionHeader } from "@/components/ui/section";
 import { useClusterStore } from "@/stores/clusterStore";
 import { useT } from "@/i18n/useT";
@@ -25,31 +16,27 @@ export function IntegrationsList() {
   const currentContext = useClusterStore((state) => state.currentContext);
 
   return (
-    <SettingsSearchProvider>
-      <div className="flex max-w-3xl flex-col gap-[18px]">
-        <SectionHeader
-          title={t("nav", "integrations")}
-          description={t("settings", "integrationsHint")}
-        />
-        {currentContext ? (
-          <SettingsSectionScope id="integrations">
-            <IntegrationsSettings />
-          </SettingsSectionScope>
-        ) : (
-          // An empty list and an unanswerable question look the same and
-          // mean different things: one says the cluster has none of these,
-          // the other that nothing was asked.
-          <div className="max-w-[64ch]">
-            <h3 className="text-xs font-medium text-fg">
-              {t("cluster", "notConnected")}
-            </h3>
-            <p className="mt-1.5 text-xs text-fg-mut">
-              {t("empty", "integrationsNoCluster")}
-            </p>
-          </div>
-        )}
-      </div>
-    </SettingsSearchProvider>
+    <div className="flex max-w-3xl flex-col gap-[18px]">
+      <SectionHeader
+        title={t("nav", "integrations")}
+        description={t("cluster", "integrationsHint")}
+      />
+      {currentContext ? (
+        <IntegrationsCatalog />
+      ) : (
+        // An empty list and an unanswerable question look the same and
+        // mean different things: one says the cluster has none of these,
+        // the other that nothing was asked.
+        <div className="max-w-[64ch]">
+          <h3 className="text-xs font-medium text-fg">
+            {t("cluster", "notConnected")}
+          </h3>
+          <p className="mt-1.5 text-xs text-fg-mut">
+            {t("empty", "integrationsNoCluster")}
+          </p>
+        </div>
+      )}
+    </div>
   );
 }
